Fix ByteBuf.ReadInt
It is not useful to write a 32-bit integer and then read it as a 32-bit float.
This is likely the long-standing result of an untested copy-paste from the code above it.

+++ main/java/com/luacraft/meta/LuaByteBuf.java
@@ -335,7 +335,7 @@ public class LuaByteBuf {
 	public static JavaFunction ReadInt = new JavaFunction() {
 		public int invoke(LuaState l) {
 			PacketBuffer self = (PacketBuffer) l.checkUserdata(1, PacketBuffer.class, "ByteBuf");
-			l.pushNumber(self.readFloat());
+			l.pushNumber(self.readInt());
 			return 1;
 		}
 	};
